Allow us to build against the C++14 standard, while not breaking compatibility with C++98.
Clang 13 and 15 on FreeBSD can successfully compile OpenOffice
when "-std=c++14" is set in gbuild and dmake, and Clang 15 can
still compile it with these changes and "-std=gnu++98".

Most of the changes involved fixing bad code, eg. NULL being
implicitly converted to sal_False, NULL being converted to typed
pointers using reinterpet_cast instead of static_cast, and integer
conversions to shorter integers.

Note that this does not change the currently targeted C++
standard on Linux and FreeBSD, which remains gnu++98.

Patch by: me

(cherry picked from commit a72d41dcd002e8c9b3d75696f0625302690e2d65)

+++ main/vcl/unx/gtk/a11y/atkwindow.cxx
@@ -291,13 +291,13 @@ ooo_window_wrapper_get_type (void)
         
         static const GTypeInfo typeInfo =
         {
-            type_query.class_size,
+            (guint16) type_query.class_size,
             (GBaseInitFunc) NULL,
             (GBaseFinalizeFunc) NULL,
             (GClassInitFunc) ooo_window_wrapper_class_init,
             (GClassFinalizeFunc) NULL,
             NULL,
-            type_query.instance_size,
+            (guint16) type_query.instance_size,
             0,
             (GInstanceInitFunc) NULL,
             NULL
